Switch to optionally importing torch (in graph translators)
