+++ gqlalchemy/transformations/translators/dgl_translator.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import List, Union
+from typing import Union
 
 import dgl
 import torch
@@ -52,22 +52,6 @@ class DGLTranslator(Translator):
         lazy: bool = MG_LAZY,
     ) -> None:
         super().__init__(host, port, username, password, encrypted, client_name, lazy)
-
-    @classmethod
-    def validate_features(cls, features: List, expected_num: int):
-        """Return true if features are okay to be set on all nodes/features.
-        Args:
-            features: To be set on all nodes. It can be anything that can be converted to torch tensor.
-            expected_num: This can be number of nodes or number of edges depending on whether features will be set on nodes or edges.
-        Returns:
-            None if features cannot be set or tensor of same features.
-        """
-        if len(features) != expected_num:
-            return None
-        try:
-            return torch.tensor(features, dtype=torch.float32)
-        except ValueError:
-            return None
 
     def to_cypher_queries(self, graph: Union[dgl.DGLGraph, dgl.DGLHeteroGraph]):
         """Produce cypher queries for data saved as part of the DGL graph. The method handles both homogeneous and heterogeneous graph. If the graph is homogeneous, a default DGL's labels will be used.
