+++ gqlalchemy/transformations/translators/translator.py
@@ -17,6 +17,11 @@ from typing import Callable, List, Set, Dict, Tuple
 from collections import defaultdict
 from numbers import Number
 
+try:
+    import torch
+except ModuleNotFoundError:
+    torch = None
+
 from gqlalchemy.transformations.constants import LABELS_CONCAT, DEFAULT_NODE_LABEL, DEFAULT_EDGE_TYPE
 from gqlalchemy.memgraph_constants import (
     MG_HOST,
@@ -91,6 +96,25 @@ class Translator(ABC):
             return len(obj) and Translator._is_most_inner_type_number(obj[0])
         # Try to parser string into number
         return isinstance(obj, Number)
+
+    @classmethod
+    def validate_features(cls, features: List, expected_num: int):
+        """Return true if features are okay to be set on all nodes/features.
+        Args:
+            features: To be set on all nodes. It can be anything that can be converted to torch tensor.
+            expected_num: This can be number of nodes or number of edges depending on whether features will be set on nodes or edges.
+        Returns:
+            None if features cannot be set or tensor of same features.
+        """
+        if torch is None:
+            raise ModuleNotFoundError("No module named 'torch'")
+
+        if len(features) != expected_num:
+            return None
+        try:
+            return torch.tensor(features, dtype=torch.float32)
+        except ValueError:
+            return None
 
     @classmethod
     def get_properties(cls, properties, entity_id):
